Add dependencyGuard Gradle plugin (#23)
Resolves #10 

Should we need to rebaseline the dependencies that we ship we must
manually run `dependencyGuardBaseline`. However we should inspect the
changes first and decide whether the changes are within our policy. We
should not just blindly accept updates.

In general we should aim for maximum compatibility and depend on the
_lowest version of dependencies that we can get away with_. That means
not updating to the latest version of a dependency just because it
happens to be released as this will force our consumers to upgrade their
project too. That's something we should avoid if possible.

Co-authored-by: Mez Pahlan <mez.pahlan@just-eat.com>